Adds Decimal mapper type
Shale Mappers generated from XSD's will use the Decimal mapper type to preserve the precision
when reading decimal values from XML documents instead of converting them to floats.

When the mapper is used for mapping other formats than XML the internal BigDecimal values are converted to floats
as they would otherwise be written as strings in output documents.

It should be possible to modify this in the future if the underlying serializers can handle BigDecimals.